Исправляет линтер, импортирует переменные в модули там, где необходимо, добавляет код в модули data и util
