+++ js/main.js
@@ -1,6 +1,6 @@
 import './util.js';
 import './data.js';
-import './functions.js';
+// import './functions.js';
 
 import {createPicture, getPictures} from './generator-id/create-picture.js';
 getPictures();
@@ -10,12 +10,8 @@ import {createMessage, createComment} from './generator-id/comment-message.js';
 createMessage();
 createComment();
 
-import {getRandomInteger, getRandomArrayElements} from './generator-id/random-integer-elements.js';
-getRandomInteger();
-getRandomArrayElements();
-
-import {createIdGenerator, generateCommentId} from './generator-id/create-id-generator.js';
+import {createIdGenerator, generateCommentId} from './generator-id/create-generator-index.js';
 createIdGenerator();
 generateCommentId();
 
-
+getPictures();
